v2/cli: do not os.Exit from the cobra command
We want to have a chance to finish tasks in case oc-mirror execution
fails, e.g., writing out the cpu and mem prof files. Doing `os.Exit`
directly means that not even deferred functions can run, so we just want
to do that at the topmost level of `main`.
